Cuvanje podataka, datum, vrijeme, mala promjena unos layouta
-Cuvanje podataka preko JSON u SharedPreference
-Klase Datum i Vrijeme umjesto Date i Time
-TrajanjeString prebacena iz main u klasu Obaveza
-Prebacio opis na dno i stavio default tacan datum i vrijeme

+++ app/src/main/java/com/example/andrej/timely/NeodredjenaObaveza.java
@@ -34,7 +34,7 @@ public class NeodredjenaObaveza extends Obaveza {
         this.visePuta = visePuta;
     }
 
-    public NeodredjenaObaveza(String naziv, boolean ispunjena, int trajanje, String komentar, Date rok, boolean visePuta) {
+    public NeodredjenaObaveza(String naziv, boolean ispunjena, float trajanje, String komentar, Date rok, boolean visePuta) {
         super(naziv, ispunjena, trajanje, komentar);
         this.rok = rok;
         this.visePuta = visePuta;
